Refactor ZWS handling by consolidating related functions into a new helpers module. Updated various extensions to utilize the new helpers, improving code organization and maintainability.
