+++ web-frontend/modules/core/components/formula/extensions/helpers.js
@@ -1,3 +1,23 @@
+// ── ZWS constants and helpers ──────────────────────────────────────
+
+export const ZWS = '\u200B'
+export const ZWS_REGEX = /\u200B/g
+export const CONSECUTIVE_ZWS_REGEX = /\u200B+/g
+
+export function isZWSNode(node) {
+  return node?.isText && node.text && /^\u200B+$/.test(node.text)
+}
+
+export function zwsTextJSON() {
+  return { type: 'text', text: ZWS }
+}
+
+export function zwsTextNode(schema) {
+  return schema.text(ZWS)
+}
+
+// ── ProseMirror document helpers ───────────────────────────────────
+
 /**
  * Returns the text content from the text node at the cursor position,
  * up to the cursor offset. Uses $pos for direct node access instead
