+++ web-frontend/modules/core/formula/tiptap/fromTipTapVisitor.js
@@ -1,4 +1,4 @@
-import { ZWS, ZWS_REGEX } from '@baserow/modules/core/components/formula/extensions/zwsHelpers'
+import { ZWS, ZWS_REGEX } from '@baserow/modules/core/components/formula/extensions/helpers'
 
 const ZWS_MARKER = Symbol('zws_marker')
